+++ rust/loopflow/src/conversation/mod.rs
@@ -1,0 +1,7 @@
+//! Shared conversation and streamed turn vocabulary.
+//!
+//! Harness drivers produce [`types::ConversationEvent`] values; the wave
+//! listener folds them into [`turns::ChatTurn`] wire frames.
+
+pub mod turns;
+pub mod types;
