+++ rust/loopflow/src/engine/repo.rs
@@ -1,0 +1,16 @@
+use anyhow::Result;
+use std::path::PathBuf;
+
+pub fn find_repo_root() -> Result<PathBuf> {
+    let output = std::process::Command::new("git")
+        .args(["rev-parse", "--show-toplevel"])
+        .output()?;
+
+    if !output.status.success() {
+        return Ok(std::env::current_dir()?);
+    }
+
+    Ok(PathBuf::from(
+        String::from_utf8_lossy(&output.stdout).trim(),
+    ))
+}
