+++ rust/loopflow/src/harness/codex_mapping.rs
@@ -4,9 +4,7 @@
 
 use serde_json::Value;
 
-use crate::lfd::conversations::types::{
-    ConversationItem, FileEdit, ItemDelta, Lifecycle, TurnUsage,
-};
+use crate::conversation::types::{ConversationItem, FileEdit, ItemDelta, Lifecycle, TurnUsage};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(super) enum ItemPhase {
