+++ rust/loopflow/src/harness/testdata/claude_crash_mid_tool.ndjson
