+++ rust/loopflow/src/harness/testdata/claude_multi_tool.ndjson
