+++ rust/loopflow/src/harness/testdata/claude_normal_turn.ndjson
