+++ rust/loopflow/src/harness/testdata/opencode_error_turn.ndjson
