+++ rust/loopflow/src/harness/testdata/opencode_normal_turn.ndjson
