+++ rust/loopflow/src/harness/testdata/opencode_tool_lifecycle.ndjson
