+++ rust/loopflow/src/harness/testdata/opencode_trace_manifest.json
