+++ rust/loopflow/src/wave/mind.rs
@@ -75,12 +75,12 @@ use chrono::{DateTime, Utc};
 use tokio::sync::mpsc;
 use tokio::time::Instant;
 
+use crate::conversation::turns::{ChatRole, ChatTurn};
+use crate::conversation::types::{ConversationEvent, ConversationItem, Lifecycle};
 use crate::engine::agent::AgentConfig;
 use crate::engine::flow::{available_flow_names, load_goal, render_goal, GoalRenderContext};
-use crate::lfd::conversations::harness::{is_terminal_harness_error, Harness};
-use crate::lfd::conversations::turns::{ChatRole, ChatTurn};
-use crate::lfd::conversations::types::{ConversationEvent, ConversationItem, Lifecycle};
-use crate::lfd::http::routes::wave_config::{read_wave_config, WaveCronDef};
+use crate::engine::wave_config::{read_wave_config, WaveCronDef};
+use crate::harness::{is_terminal_harness_error, Harness};
 use crate::wave::journal::{ellipsize, MessageId, MessageOp, PendingMessage};
 use crate::wave::memory::Memory;
 use crate::wave::resident::ListenerClient;
@@ -1113,8 +1113,8 @@ mod tests {
 
     use async_trait::async_trait;
 
-    use crate::lfd::conversations::harness::Capabilities;
-    use crate::lfd::conversations::types::TurnUsage;
+    use crate::conversation::types::TurnUsage;
+    use crate::harness::Capabilities;
     use crate::wave::journal::{journal_path, EventKind, Journal};
     use crate::wave::runtime::WaveRuntime;
     use crate::wave::server::{self, ResidentDoor};
