+++ rust/loopflow/src/wave/wire.rs
@@ -37,7 +37,7 @@
 
 use serde::{Deserialize, Serialize};
 
-use crate::lfd::conversations::types::{ConversationItem, Lifecycle};
+use crate::conversation::types::{ConversationItem, Lifecycle};
 use crate::wave::journal::{Attribution, MessageOp};
 
 /// Header carrying the resident token on every `/resident/*` request.
